fix(auth): address header review feedback
