+++ src/components/app/index.ts
@@ -1,2 +1,3 @@
 export { AppToasterProvider } from "./AppToaster";
 export { AppShellHeader } from "./AppShellHeader";
+export type { AuthActionState } from "./AppShellHeader";
