+++ src/features/profiles/utils/profilePresentation.ts
@@ -1,3 +1,5 @@
+import { getProfileAvatarFallbackLabel } from "@/lib/profilePresentation";
+
 import { ProfileDataAccessError } from "../queries/profileApi";
 
 type ProfileLoadErrorCopy = {
@@ -51,19 +53,4 @@ export function getProfileLoadErrorCopy(error: unknown): ProfileLoadErrorCopy {
   };
 }
 
-export function getProfileAvatarFallbackLabel(displayName: string): string {
-  const trimmedName = displayName.trim();
-
-  if (trimmedName === "") {
-    return "RB";
-  }
-
-  const initials = trimmedName
-    .split(/\s+/)
-    .filter((token) => token !== "")
-    .slice(0, 2)
-    .map((token) => token[0]?.toUpperCase() ?? "")
-    .join("");
-
-  return initials === "" ? "RB" : initials;
-}
+export { getProfileAvatarFallbackLabel };
